+++ durapy/src/unipy/uniops/__init__.py
@@ -3,4 +3,12 @@ The `DuraPy` `UniOps` module for Control and Operations tasks.
 `UniOps` is built for tasks related to mechatronics control and operation tasks.
 """
 
-__all__ = ["uniops.py"]
+from conpidcon import (ContinuousPIDController,)
+from forward_kinematics import ()
+from inverse_kinematics import ()
+
+__all__ = [
+    "conpidcon.py", 
+    "forward_kinematics.py", 
+    "inverse_kinematics.py"
+]