+++ durapy/src/unipy/uniops/inverse_kinematics.py
@@ -1,0 +1,7 @@
+"""
+The `DuraPy` `UniOps` Inverse Kinematics module.
+
+This module contains resources and tools for inverse kinematics calculations.
+"""
+
+from durapy.src.unipy.unimath import linalg_dtypes